backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-multipart
transformers
scikit-learn
torch
numpy
python-dotenv
requests
